Pre: Move Entid and KnownEntid into core_traits
