+++ core/Cargo.toml
@@ -14,6 +14,9 @@ uuid = { version = "0.5", features = ["v4", "serde"] }
 serde = { version = "1.0", features = ["rc"] }
 serde_derive = "1.0"
 
+[dependencies.core_traits]
+path = "../core-traits"
+
 [dependencies.edn]
 path = "../edn"
 features = ["serde_support"]
