+++ db/Cargo.toml
@@ -32,6 +32,9 @@ path = "../edn"
 [dependencies.mentat_core]
 path = "../core"
 
+[dependencies.core_traits]
+path = "../core-traits"
+
 [dependencies.mentat_sql]
 path = "../sql"
