+++ db/src/tx_checking.rs
@@ -13,8 +13,11 @@ use std::collections::{
     BTreeMap,
 };
 
-use mentat_core::{
+use core_traits::{
     Entid,
+};
+
+use mentat_core::{
     TypedValue,
     ValueType,
 };
