+++ query-algebrizer/Cargo.toml
@@ -10,6 +10,9 @@ failure_derive = "0.1.1"
 [dependencies.mentat_core]
 path = "../core"
 
+[dependencies.core_traits]
+path = "../core-traits"
+
 [dependencies.mentat_query]
 path = "../query"
