+++ query-algebrizer/src/clauses/pattern.rs
@@ -8,9 +8,12 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
+use core_traits::{
+    Entid,
+};
+
 use mentat_core::{
     Cloned,
-    Entid,
     HasSchema,
     TypedValue,
     ValueType,
