+++ query-algebrizer/src/types.rs
@@ -14,8 +14,11 @@ use std::fmt::{
     Formatter,
 };
 
-use mentat_core::{
+use core_traits::{
     Entid,
+};
+
+use mentat_core::{
     TypedValue,
     ValueRc,
     ValueType,
