+++ query-projector/Cargo.toml
@@ -12,6 +12,9 @@ indexmap = "1"
 version = "0.13"
 features = ["limits"]
 
+[dependencies.core_traits]
+path = "../core-traits"
+
 [dependencies.mentat_core]
 path = "../core"
