+++ query-projector/src/lib.rs
@@ -16,6 +16,7 @@ extern crate indexmap;
 extern crate rusqlite;
 
 extern crate mentat_core;
+extern crate core_traits;
 extern crate mentat_db;                 // For value conversion.
 extern crate mentat_query;
 extern crate mentat_query_algebrizer;
