+++ query-projector/tests/aggregates.rs
@@ -9,13 +9,17 @@
 // specific language governing permissions and limitations under the License.
 
 extern crate mentat_core;
+extern crate core_traits;
 extern crate mentat_query;
 extern crate mentat_query_algebrizer;
 extern crate mentat_query_projector;
 
+use core_traits::{
+    Entid,
+};
+
 use mentat_core::{
     Attribute,
-    Entid,
     Schema,
     ValueType,
 };
