+++ query-pull/Cargo.toml
@@ -14,6 +14,9 @@ features = ["limits"]
 [dependencies.mentat_core]
 path = "../core"
 
+[dependencies.core_traits]
+path = "../core-traits"
+
 [dependencies.mentat_db]
 path = "../db"
