+++ src/store.rs
@@ -26,8 +26,11 @@ use rusqlite;
 
 use edn;
 
-use mentat_core::{
+use core_traits::{
     Entid,
+};
+
+use mentat_core::{
     Keyword,
     StructuredMap,
     TxReport,
